xz_wrapper{,_extended}: Detect if liblzma lacks support for a BCJ filter
This way a useful error message is printed if user has requested a filter
that the liblzma version in use doesn't support. This matters with the
ARM64 and RISC-V filters if Squashfs-tools is running with an liblzma
version. (In theory it can matter also if using a custom liblzma build
that has some filters disabled, but such builds aren't for general use.)

Signed-off-by: Lasse Collin <lasse.collin@tukaani.org>